+++ transport-native-io_uring/src/test/java/io/netty/channel/uring/UserDataTest.java
@@ -27,7 +27,7 @@ public class UserDataTest {
             for (int op = 0; op < 20; op++) {
                 for (int data = Short.MIN_VALUE; data <= Short.MAX_VALUE; data++) {
                     final int expectedFd = fd;
-                    final int expectedOp = op;
+                    final byte expectedOp = (byte) op;
                     final short expectedData = (short) data;
                     long udata = UserData.encode(expectedFd, expectedOp, expectedData);
                     UserData.decode(0, 0, udata, new IOUringCompletionQueueCallback() {
